# openapi_spec.yaml
openapi: 3.0.3
info:
  title: Docaiche API
  description: |
    AI-powered documentation search and content management system with advanced analytics capabilities.
    
    ## Overview
    Docaiche provides intelligent documentation search, content enrichment, and comprehensive analytics
    for tracking user behavior and system performance. The API supports real-time search with caching,
    content ingestion from multiple sources, and AI-powered content enhancement.
    
    ## Key Features
    - **Intelligent Search**: AI-powered search with technology filtering and relevance scoring
    - **Content Management**: Ingest and manage documentation from various sources
    - **Analytics Dashboard**: Comprehensive metrics for search patterns, user behavior, and system performance
    - **AI Enrichment**: Automatic content enhancement using LLM providers
    - **Real-time Monitoring**: Live system health and performance metrics
    
    ## Analytics Capabilities
    The analytics endpoints provide rich data for building beautiful dashboards:
    - Search query trends and patterns over time
    - User engagement metrics (clicks, dwell time, scroll depth)
    - Content quality scores and performance
    - System performance metrics (response times, cache hit rates)
    - Technology adoption trends
    - Real-time activity feeds
    
    ## Monitoring Integration
    - **Loki**: Centralized log aggregation for all services
...
    
  version: 2.0.0
  contact:
...
  license:
    name: MIT
...
servers:
...
paths:
...
  /analytics:
    get:
      tags: [analytics]
      summary: Detailed analytics data
      description: |
        Returns comprehensive analytics for the specified time range.
        
        **Dashboard Visualization Ideas:**
        - Search query word clouds from top queries
        - Technology adoption pie charts
        - User engagement funnel charts
        - Time series graphs for all metrics
        - Geographic distribution maps (if user location data available)
        
        **Data Granularity:**
        - Supports 24h, 7d, and 30d time ranges
        - Returns aggregated metrics suitable for visualization
        - Includes top-N lists for queries and technologies
        
        **Current Status:** Returns demonstration data showing the structure
        and types of analytics available. Will return real data as the system
        accumulates usage metrics.
      operationId: getAnalytics
      parameters:
        - name: timeRange
          in: query
          description: Time range for analytics
          required: false
          schema:
            type: string
            enum: [24h, 7d, 30d]
            default: 24h
      responses:
        '200':
          description: Analytics data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsResponse'
...
  /enrichment/enrich:
    post:
      tags: [enrichment]
      summary: Enrich content
      description: |
        Trigger AI-powered content enrichment.
        
        **Enrichment Capabilities:**
        - Generate missing documentation
        - Improve content quality
        - Add code examples
        - Create summaries
        - Fix technical errors
        
        **Currently:** Returns 503 as service is not initialized
      operationId: enrichContent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnrichmentRequest'
      responses:
        '202':
          description: Enrichment started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnrichmentResponse'
        '503':
          description: Enrichment service unavailable
components:
  schemas:
...
    AnalyticsResponse:
      type: object
      required: [timeRange, searchMetrics, contentMetrics, userMetrics, timestamp]
      properties:
        timeRange:
          type: string
          enum: [24h, 7d, 30d]
        searchMetrics:
          type: object
          properties:
            totalSearches:
              type: integer
              description: Total searches in period
            avgResponseTime:
              type: number
              description: Average response time
            successRate:
              type: number
              description: Search success rate (0-1)
            topQueries:
              type: array
              items:
                type: object
                properties:
                  query:
                    type: string
                  count:
                    type: integer
        contentMetrics:
          type: object
          properties:
            totalDocuments:
              type: integer
            totalChunks:
              type: integer
            avgQualityScore:
              type: number
            documentsByTechnology:
              type: array
              items:
                type: object
                properties:
                  technology:
                    type: string
                  count:
                    type: integer
        userMetrics:
          type: object
          properties:
            activeUsers:
              type: integer
            totalSessions:
              type: integer
            avgSessionDuration:
              type: number
              description: Average session duration in milliseconds
        timestamp:
          type: string
          format: date-time
...
    EnrichmentRequest:
      type: object
      required: [content_id]
      properties:
        content_id:
          type: string
          description: Document to enrich
        enrichment_type:
          type: string
          enum: [quality_improvement, gap_filling, example_generation, summary]
        target_quality_score:
          type: number
          minimum: 0
          maximum: 1
    EnrichmentResponse:
      type: object
      required: [enrichment_id, status]
      properties:
        enrichment_id:
          type: string
        status:
          type: string
          enum: [queued, processing, completed, failed]
        estimated_completion:
          type: string
          format: date-time
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication (future implementation)
security: []  # No security currently implemented